Split ThreadContext extensions and simplify lifecycle management
