+++ log4j-api-test/src/main/java/org/apache/logging/log4j/test/junit/UsingThreadContextStack.java
@@ -37,7 +37,7 @@ import java.lang.annotation.Target;
 @Target({ElementType.TYPE, ElementType.METHOD, ElementType.ANNOTATION_TYPE})
 @Documented
 @Inherited
-@ExtendWith(ThreadContextExtension.class)
+@ExtendWith(ThreadContextStackExtension.class)
 @ReadsSystemProperty
 public @interface UsingThreadContextStack {
 }
